+++ NEWS.md
@@ -10,6 +10,9 @@ shiny 1.5.0.9000
 
 * Fixed #2859: `renderPlot()` wasn't correctly setting `showtext::showtext_opts()`'s `dpi` setting with the correct resolution on high resolution displays; which means, if the font was rendered by showtext, font sizes would look smaller than they should on such displays. (#2941)
 
+### Library updates
+
+* Removed html5shiv and respond.js, which were used for IE 8 and IE 9 compatibility. (#2973)
 
 shiny 1.5.0
 ===========